# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TalentHub API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: jobs
    description: Job listing operations
  - name: applications
    description: Job application operations
  - name: stats
    description: Dashboard statistics
  - name: auth
    description: Authentication
  - name: companies
    description: Company management
  - name: users
    description: User management
  - name: storage
    description: File storage
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all job postings with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: jobType
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: salaryMin
          in: query
          schema:
            type: integer
        - name: salaryMax
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            description: newest, oldest, salary-high, salary-low, most-applied
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a new job listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a single job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /jobs/{id}/view:
    post:
      operationId: incrementJobView
      tags: [jobs]
      summary: Increment view count for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: View count incremented
          content:
            application/json:
              schema:
                type: object
                required:
                  - viewCount
                properties:
                  viewCount:
                    type: integer

  /jobs/{id}/applications:
    get:
      operationId: listJobApplications
      tags: [applications]
      summary: List applications for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: applyToJob
      tags: [applications]
      summary: Submit an application for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications, optionally filtered by applicant email
      parameters:
        - name: email
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Applications with job details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationWithJob"

  /applications/{id}/status:
    patch:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Update the status of an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats]
      summary: Get high-level job board stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"

  /stats/by-category:
    get:
      operationId: getStatsByCategory
      tags: [stats]
      summary: Job counts grouped by category
      responses:
        "200":
          description: Stats by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryStat"

  /stats/recent-jobs:
    get:
      operationId: getRecentJobs
      tags: [stats]
      summary: Most recently posted jobs
      responses:
        "200":
          description: Recent jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"

  /stats/applications-over-time:
    get:
      operationId: getApplicationsOverTime
      tags: [stats]
      summary: Application counts per day for last 30 days
      responses:
        "200":
          description: Applications over time
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyStat"

  /auth/user/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new candidate account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "409":
          description: Email already in use

  /auth/user/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login as a candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "401":
          description: Invalid credentials

  /auth/user/logout:
    post:
      operationId: logoutUser
      tags: [auth]
      summary: Logout candidate
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /auth/user/me:
    get:
      operationId: getUserMe
      tags: [auth]
      summary: Get current candidate session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "401":
          description: Not authenticated
    patch:
      operationId: updateUserProfile
      tags: [auth]
      summary: Update candidate profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"

  /auth/company/register:
    post:
      operationId: registerCompany
      tags: [auth]
      summary: Register a new company account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyRegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthCompanyResponse"
        "409":
          description: Email already in use

  /auth/company/login:
    post:
      operationId: loginCompany
      tags: [auth]
      summary: Login as a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthCompanyResponse"
        "401":
          description: Invalid credentials

  /auth/company/logout:
    post:
      operationId: logoutCompany
      tags: [auth]
      summary: Logout company
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /auth/company/me:
    get:
      operationId: getCompanyMe
      tags: [auth]
      summary: Get current company session
      responses:
        "200":
          description: Current company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthCompanyResponse"
        "401":
          description: Not authenticated
    patch:
      operationId: updateCompanyProfile
      tags: [auth]
      summary: Update company profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthCompanyResponse"

  /admin/companies:
    get:
      operationId: adminListCompanies
      tags: [companies]
      summary: Admin - list all companies
      responses:
        "200":
          description: Companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyPublic"

  /admin/companies/{id}:
    patch:
      operationId: adminUpdateCompany
      tags: [companies]
      summary: Admin - approve or update a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCompanyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyPublic"
    delete:
      operationId: adminDeleteCompany
      tags: [companies]
      summary: Admin - delete a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /admin/users:
    get:
      operationId: adminListUsers
      tags: [users]
      summary: Admin - list all users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPublic"

  /admin/users/{id}:
    delete:
      operationId: adminDeleteUser
      tags: [users]
      summary: Admin - delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted


components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Job:
      type: object
      required:
        - id
        - title
        - company
        - location
        - jobType
        - category
        - description
        - salaryMin
        - salaryMax
        - featured
        - createdAt
        - applicationCount
        - viewCount
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        companyLogo:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        location:
          type: string
        remote:
          type: boolean
        jobType:
          type: string
          description: full-time, part-time, contract, internship
        category:
          type: string
          description: Engineering, Design, Marketing, Finance, etc.
        description:
          type: string
        requirements:
          type: ["string", "null"]
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        currency:
          type: string
        featured:
          type: boolean
        applicationCount:
          type: integer
        viewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time

    JobInput:
      type: object
      required:
        - title
        - company
        - location
        - jobType
        - category
        - description
        - salaryMin
        - salaryMax
      properties:
        title:
          type: string
          minLength: 1
        company:
          type: string
          minLength: 1
        companyLogo:
          type: string
        companyId:
          type: integer
        location:
          type: string
          minLength: 1
        remote:
          type: boolean
        jobType:
          type: string
        category:
          type: string
        description:
          type: string
          minLength: 1
        requirements:
          type: string
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        currency:
          type: string
        featured:
          type: boolean
        expiresAt:
          type: string
          format: date-time

    JobUpdate:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        companyLogo:
          type: string
        companyId:
          type: integer
        location:
          type: string
        remote:
          type: boolean
        jobType:
          type: string
        category:
          type: string
        description:
          type: string
        requirements:
          type: string
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        currency:
          type: string
        featured:
          type: boolean
        expiresAt:
          type: string
          format: date-time

    Application:
      type: object
      required:
        - id
        - jobId
        - applicantName
        - applicantEmail
        - status
        - createdAt
      properties:
        id:
          type: integer
        jobId:
          type: integer
        applicantName:
          type: string
        applicantEmail:
          type: string
        phone:
          type: ["string", "null"]
        coverLetter:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        status:
          type: string
          description: pending, reviewed, shortlisted, rejected
        createdAt:
          type: string
          format: date-time

    ApplicationWithJob:
      type: object
      required:
        - id
        - jobId
        - applicantName
        - applicantEmail
        - status
        - createdAt
        - jobTitle
        - company
        - location
        - jobType
      properties:
        id:
          type: integer
        jobId:
          type: integer
        applicantName:
          type: string
        applicantEmail:
          type: string
        phone:
          type: ["string", "null"]
        coverLetter:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        jobTitle:
          type: string
        company:
          type: string
        location:
          type: string
        jobType:
          type: string
        companyLogo:
          type: ["string", "null"]

    ApplicationInput:
      type: object
      required:
        - applicantName
        - applicantEmail
      properties:
        applicantName:
          type: string
          minLength: 1
        applicantEmail:
          type: string
          format: email
        phone:
          type: string
        coverLetter:
          type: string
        resumeUrl:
          type: string
        profileImageUrl:
          type: string

    ApplicationStatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          description: pending, reviewed, shortlisted, rejected

    StatsSummary:
      type: object
      required:
        - totalJobs
        - totalApplications
        - featuredJobs
        - newJobsThisWeek
      properties:
        totalJobs:
          type: integer
        totalApplications:
          type: integer
        featuredJobs:
          type: integer
        newJobsThisWeek:
          type: integer

    CategoryStat:
      type: object
      required:
        - category
        - count
      properties:
        category:
          type: string
        count:
          type: integer

    DailyStat:
      type: object
      required:
        - date
        - count
      properties:
        date:
          type: string
        count:
          type: integer

    LoginInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    UserRegisterInput:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string
        location:
          type: string

    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        bio:
          type: string
        location:
          type: string
        profileImageUrl:
          type: string
        resumeUrl:
          type: string

    AuthUserResponse:
      type: object
      required:
        - id
        - name
        - email
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    UserPublic:
      type: object
      required:
        - id
        - name
        - email
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    CompanyRegisterInput:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        website:
          type: string
        industry:
          type: string
        description:
          type: string
        size:
          type: string
        location:
          type: string

    CompanyProfileUpdate:
      type: object
      properties:
        name:
          type: string
        website:
          type: string
        industry:
          type: string
        description:
          type: string
        size:
          type: string
        location:
          type: string
        logoUrl:
          type: string

    AuthCompanyResponse:
      type: object
      required:
        - id
        - name
        - email
        - approved
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        logoUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        approved:
          type: boolean
        createdAt:
          type: string
          format: date-time

    CompanyPublic:
      type: object
      required:
        - id
        - name
        - email
        - approved
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        logoUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        approved:
          type: boolean
        createdAt:
          type: string
          format: date-time

    AdminCompanyUpdate:
      type: object
      properties:
        approved:
          type: boolean
        name:
          type: string

    OkResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean