# lib/api-zod/src/generated/types/adminCompanyUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TalentHub API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminCompanyUpdate {
  approved?: boolean;
  name?: string;
}