# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TalentHub API
 * OpenAPI spec version: 0.1.0
 */

export interface ApplicationInput {
  /** @minLength 1 */
  applicantName: string;
  applicantEmail: string;
  phone?: string;
  coverLetter?: string;
  resumeUrl?: string;
  profileImageUrl?: string;
}