# lib/api-zod/src/generated/types/applicationWithJob.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TalentHub API
 * OpenAPI spec version: 0.1.0
 */

export interface ApplicationWithJob {
  id: number;
  jobId: number;
  applicantName: string;
  applicantEmail: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  status: string;
  createdAt: Date;
  jobTitle: string;
  company: string;
  location: string;
  jobType: string;
  /** @nullable */
  companyLogo?: string | null;
}