# lib/api-zod/src/generated/types/authCompanyResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TalentHub API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthCompanyResponse {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  location?: string | null;
  approved: boolean;
  createdAt: Date;
}