# lib/api-zod/src/generated/types/authUserResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TalentHub API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUserResponse {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  createdAt: Date;
}