# lib/api-zod/src/generated/types/companyProfileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TalentHub API
 * OpenAPI spec version: 0.1.0
 */

export interface CompanyProfileUpdate {
  name?: string;
  website?: string;
  industry?: string;
  description?: string;
  size?: string;
  location?: string;
  logoUrl?: string;
}